requirements.txt
fastapi~=0.115.3
SQLAlchemy~=2.0.36
pydantic~=2.9.2
uvicorn~=0.32.0
python-dotenv~=1.0.1
starlette~=0.41.0
PyJWT~=2.10.1
passlib~=1.7.4